refactor: move formatting to utils
